fix: explicitly typyecast to int for ranint input

+++ analytics_data_api/management/commands/generate_data.py
@@ -236,10 +236,10 @@ def generate_learner_engagement_data(course_id, username, start_date, end_date, 
     while current < end_date:
         current = current + datetime.timedelta(days=1)
         for metric in engagement_events.INDIVIDUAL_EVENTS:
-            num_events = random.randint(0, max_value)
+            num_events = random.randint(0, int(max_value))
             if num_events:
                 for _ in range(num_events):
-                    count = random.randint(0, max_value / 20)
+                    count = random.randint(0, int(max_value / 20))
                     entity_type = metric.split('_', 1)[0]
                     event = metric.split('_', 1)[1]
                     entity_id = f'an-id-{entity_type}-{event}'
